Incorporated feedback and added XRPF check
