--- Runtime/Components/OcclusionEvent.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a5d3131cce9da174ea02d05474f7fa70
-timeCreated: 1556053757
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
